# src/main/java/com/anz/common/api/AccountResponse.java
package com.anz.common.api;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

import javax.validation.Valid;
import javax.validation.constraints.NotEmpty;
import java.util.List;

@Builder
@Getter
@NoArgsConstructor
@AllArgsConstructor
public class AccountResponse {
    @NotEmpty
    List<@Valid Account> accounts;
}