# src/main/openapi/account_service.swagger.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/anz/wholesale",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/users/{user-id}/accounts": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Account list",
        "description": "Provide list of account details for requested user<br/> ",
        "operationId": "getAccountsByUserId",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account list",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Account service encountered a request validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account service is unable find requested resource",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Account service encountered an unexpected internal error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user-id}/accounts/{account-id}/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Account transaction list",
        "description": "Provide list of account transactions for requested account<br/> ",
        "operationId": "getTransactionsByAccountId",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account transaction list",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Account service encountered a request validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account service is unable find requested resource",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Account service encountered an unexpected internal error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "required": [
          "accountName",
          "accountType",
          "balanceDate",
          "accountNumber",
          "currency",
          "openingAvailableBalance"
        ],
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string"
          },
          "accountName": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "Savings",
              "Current"
            ]
          },
          "balanceDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string"
          },
          "openingAvailableBalance": {
            "$ref": "#/components/schemas/Amount"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "AccountResponse": {
        "required": [
          "accounts"
        ],
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          }
        }
      },
      "Amount": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number"
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "rel": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "media": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "string",
            "enum": [
              "200 OK",
              "400 BAD_REQUEST",
              "404 NOT_FOUND",
              "500 INTERNAL_SERVER_ERROR"
            ]
          },
          "errorId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}